MAJOR work! Authentication is now taking place in app!
Needs to be tested on Windows as portable but it's coming along!

App stores details in default location per https://medium.com/@ccnokes/how-to-store-user-data-in-electron-3ba6bf66bc1e, and GitHub API is now fully in use to check/display user info as well as check for total items in policy column (to-be-fleshed-out-next).