#2: added Carlito as default font
It's metric compatible with Calibri, which is non-free (but the suggested font from the state Style Guide). Just to avoid having to worry about users who may not have the same version as I (a possibility with machines running different forms of Office and Windows), I'm embedding Carlito instead of relying on Calibri.

I also updated a few comments throughout the code.